pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "masters"
version = "0.0.1"
description = "Masters study codebase"
requires-python = ">= 3.8, < 3.12"
dependencies = [
    "torch ~= 2.1",
    "lightning[pytorch-extra] ~= 2.1",
    "monai[nibabel, skimage, pillow, tensorboard, torchvision, itk, tqdm] ~= 1.3",
    "torchmetrics ~= 1.2",
    "torchvision ~= 0.16",
    "scipy ~= 1.10",
    "numpy ~= 1.24",
    "rich ~= 13.7",
    "pqdm ~= 0.2",
    "hydra-core ~= 1.3",
    "hydra-colorlog ~= 1.2",
    "hydra-optuna-sweeper ~= 1.2",
]
[project.urls]
"Source" = "https://github.com/WojciechCebula/masters"


[project.optional-dependencies]
jupyter = [
    "jupyterlab~=3.6",
    "pandas ~= 2.0",
    "matplotlib ~= 3.7",
    "seaborn ~= 0.12",
    "plotly ~= 5.14",
]
lint = ["ruff ~= 0.1", "pre-commit ~= 2.20"]
test = [
    "pytest ~= 7.1",
    "pytest-cases ~= 3.6",
    "pytest-cov ~= 3.0",
    "pytest-xdist ~= 2.5",
    "pytest-sugar ~= 0.9",
]
dev = ["masters[jupyter, lint, test]"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
    # "UP",  # pyupgrade
]
line-length = 120
ignore = [
    # "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "W191", # indentation contains tabs
]
extend-exclude = [".venv/", "build/", "dist/", "docs/"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint.isort]
order-by-type = true
lines-between-types = 1
combine-as-imports = true
known-local-folder = ["masters"]
